feat: drop support for GitHub Enterprise Server v3.7, v3.8, v3.9 and v3.10
BREAKING CHANGE: This drops support for exports from GitHub
Enterprise Server 3.7, 3.8, 3.9 and 3.10 as these versions have now
been [deprecated](https://docs.github.com/en/enterprise-server/admin/all-releases)
by GitHub. v3.11 is now the earliest supported version.
