+++ src/auditors/repository-actions-variables.ts
@@ -1,23 +1,13 @@
-import semver from 'semver';
-
 import { AuditorFunction, AuditorWarning } from '../types';
 import { pluralize } from '../utils';
 
 export const TYPE = 'repository-actions-variables';
 
 export const auditor: AuditorFunction = async ({
-  gitHubEnterpriseServerVersion,
   octokit,
   owner,
   repo,
 }): Promise<AuditorWarning[]> => {
-  if (
-    typeof gitHubEnterpriseServerVersion !== 'undefined' &&
-    semver.lt(gitHubEnterpriseServerVersion, '3.8.0')
-  ) {
-    return [];
-  }
-
   const { data } = await octokit.rest.actions.listRepoVariables({
     owner,
     repo,
