+++ test/auditors/repository-actions-variables.test.ts
@@ -43,30 +43,4 @@ describe('repositoryActionVariables', () => {
 
     expect(warnings).toEqual([]);
   });
-
-  it('no-ops if running against GitHub Enterprise Server <3.8', async () => {
-    const auditorArguments = buildAuditorArguments({
-      gitHubEnterpriseServerVersion: '3.7.3',
-    });
-    const warnings = await auditor(auditorArguments);
-
-    expect(warnings.length).toEqual(0);
-  });
-
-  it('runs if running against GitHub Enterprise Server >=3.8', async () => {
-    const fetch = fetchMock
-      .sandbox()
-      .getOnce('https://api.github.com/repos/test/test/actions/variables?per_page=1', {
-        total_count: 0,
-        variables: [],
-      });
-
-    const auditorArguments = buildAuditorArguments({
-      fetchMock: fetch,
-      gitHubEnterpriseServerVersion: '3.8.0',
-    });
-    const warnings = await auditor(auditorArguments);
-
-    expect(warnings.length).toEqual(0);
-  });
 });
