+++ test/auditors/repository-dependabot-alerts.test.ts
@@ -48,30 +48,4 @@ describe('repositoryDependabotAlerts', () => {
 
     expect(warnings).toEqual([]);
   });
-
-  it('no-ops if running against GitHub Enterprise Server <3.8', async () => {
-    const auditorArguments = buildAuditorArguments({
-      gitHubEnterpriseServerVersion: '3.7.3',
-    });
-    const warnings = await auditor(auditorArguments);
-
-    expect(warnings.length).toEqual(0);
-  });
-
-  it('runs if running against GitHub Enterprise Server >=3.8', async () => {
-    const fetch = fetchMock
-      .sandbox()
-      .getOnce('https://api.github.com/repos/test/test/dependabot/alerts?per_page=1', {
-        total_count: 0,
-        variables: [],
-      });
-
-    const auditorArguments = buildAuditorArguments({
-      fetchMock: fetch,
-      gitHubEnterpriseServerVersion: '3.8.0',
-    });
-    const warnings = await auditor(auditorArguments);
-
-    expect(warnings.length).toEqual(0);
-  });
 });
